requirements.txt
fastapi[standard]
pytest
sqlalchemy 
pydantic
psycopg2-binary
asyncpg
python-dotenv
python-multipart
httpx
passlib
PyJWT
uvicorn
requests